requirements.txt
aiocontextvars==0.2.2
allure-pytest==2.8.6
allure-python-commons==2.8.6
appdirs==1.4.4
Appium-Python-Client==1.2.0
atomicwrites==1.4.0
attrs==21.2.0
black==19.10b0
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
ConfigArgParse==1.5.2
contextvars==2.4
dataclasses==0.8
Flask==2.0.1
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.8.0
geventhttpclient==1.5.1
geventhttpclient-wheels==1.3.1.dev3
greenlet==1.1.1
httprunner==3.1.5
idna==3.2
immutables==0.16
importlib-metadata==4.8.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.9.5
jsonpath==0.82
locust==2.5.2.dev8
locustio==1.0b2
loguru==0.4.1
MarkupSafe==2.0.1
more-itertools==8.8.0
msgpack==1.0.2
packaging==21.0
parameterized==0.8.1
pathspec==0.9.0
pluggy==0.13.1
psutil==5.8.0
py==1.10.0
pycparser==2.20
pydantic==1.8.2
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==5.4.3
pytest-base-url==1.4.2
pytest-html==2.1.1
pytest-metadata==1.11.0
pywin32==301
PyYAML==5.4.1
pyzmq==22.3.0
regex==2021.8.28
requests==2.26.0
roundrobin==0.0.2
selenium==3.141.0
sentry-sdk==0.14.4
six==1.16.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
win32-setctime==1.0.3
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
